Add "catch" command
Catch will attempt to catch a wild pokemon. The higher the base EXP is
the harder it is to catch it. The catch is sampled 100 times to give
less based EXP pokemons more chances to get caught.

After a successful attempt, the pokemon is added to the Pokedex. The
Pokedex being inside internal API code and has to initialized in main is
a misdesign, but it's too late for me to fix it now.